some more documentation work
the todo list is now  much better and alot of what im up to with this
box is in it. its not ordered in a partculat fashion and alot of the
items aren't really high on my list so it will probably stay quite long
for some time. also i linked a crappy pic of the early hardware on the
start page and fixed some typos in the main docs.