touchscreen input patches
this patch contains all the stuff needed for my chinavasion touchscreen
it takes care of swapping all the ABS_X/Y constants in the userspace code
with ABS_Z/Rx constants.

This is a quick and dirty workaround for a problem that could most
likely also be solved in the kernel/usbhid. With this patch im at a
state where the build should run through and make something quite like
the dev firmware i have on my device right now.